+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+* Support for time goals via optional package `chronometrist-targets`.
 ### Fixed
 * Refresh buffer when clocking in (instead of waiting for first timer refresh)
 * Insertion of values is now slightly smarter about detecting and handling Lisp data types.