+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+* `chronometrist-toggle-task-no-reason` (which did nothing since the migration from timeclock.el) is now called `chronometrist-toggle-task-no-hooks`. It will toggle the task without running the before-in/after-in/before-out/after-out functions.
+### Fixed
+* Refresh buffer when clocking in (instead of waiting for first timer refresh)
+* Insertion of values is now slightly smarter about detecting and handling Lisp data types.
+
 ## [0.4.2] - 2020-01-15
 ### Fixed
 * Library headers for MELPA release